requirements.txt
gym==0.21.0
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.2
PyYAML==6.0
seaborn==0.11.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.64.0
ujson==5.7.0
